% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcwm.R
\name{conf}
\alias{conf}
\title{Confusion Matrix function to plot the Confusion Matrix
The function plots the confusion matrix of the classification results}
\usage{
conf(x)
}
\arguments{
\item{x}{The predicted classes returned from the clustering algorithm}
}
\description{
Confusion Matrix function to plot the Confusion Matrix
The function plots the confusion matrix of the classification results
}
\keyword{confusion}
\keyword{matrix}
