% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcwm.R
\name{logq}
\alias{logq}
\title{Logit Function for depencies
The function performs a logit function with centered around the maximum to ignore underflow errors}
\usage{
logq(x)
}
\arguments{
\item{x}{numeric values to be transformed between 0 and 1}
}
\description{
Logit Function for depencies
The function performs a logit function with centered around the maximum to ignore underflow errors
}
\keyword{logit}
