% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcwm.R
\name{MultiCwm}
\alias{MultiCwm}
\title{MultiCwm Function for the clustering of data}
\usage{
MultiCwm(
  Y,
  x,
  G = 1,
  init = c("random", "mclust", "kmeans"),
  maxit = 1000,
  tol = 1e-10,
  show_table = FALSE
)
}
\arguments{
\item{Y}{The categorical variables which can be any of the form: vector, one column matrix,
or a one-encoding data frame or matrix}

\item{x}{The feature or independent variable of the data.}

\item{G}{The number of components which must be an integer}

\item{init}{This is the initial value to start the clustering algorithm.
This can any of the options: random, mclust, or kmeans}

\item{maxit}{The maximum number of iterations to terminate the algorithm. This must be an integer.}

\item{tol}{This is the tolenrance level of convergence dafault to 1e-10.}

\item{show_table}{This is a logical parameter which is default to FALSE.}
}
\value{
res which is a list containing the component mean and covariance matrix, the information criterion
}
\description{
MultiCwm Function for the clustering of data
}
\keyword{Algorithm}
\keyword{Clustering}
\keyword{MultiCwm}
