% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcwm.R
\name{simData}
\alias{simData}
\title{Simulation Function for fake data
Simulated function simulates from the distributions assumed by the model}
\usage{
simData(n, G, d = 2, J, mean, sigma, w, Pi = NULL)
}
\arguments{
\item{n}{The number of values to simulate. n must be an integer}

\item{G}{The number of components which must be an integer}

\item{d}{The dimension number of features x: must be an integer.}

\item{J}{The dimension number of categorical variable Y: must be an integer}

\item{mean}{The true parameter mean for each component on each row of the matrix}

\item{sigma}{The true values of the variance-covariance matrix}

\item{w}{The true values of coefficients for the multinomial regression Y ~ X}

\item{Pi}{The mixing probability of the components}
}
\description{
Simulation Function for fake data
Simulated function simulates from the distributions assumed by the model
}
\keyword{data}
\keyword{simulated}
