% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcwm.R
\name{zipCWM}
\alias{zipCWM}
\title{zipCWM Function for the clustering of data}
\usage{
zipCWM(
  Y,
  w,
  u,
  v,
  G = 1,
  maxit = 1000,
  tol = 1e-10,
  show_table = FALSE
)
}
\arguments{
\item{Y}{The count variables which can be any of the form: vector, one column matrix}

\item{v}{The feature or independent variable of the data.}

\item{u}{The feature or independent variable of the data.}

\item{w}{The feature or independent variable of the data.}

\item{G}{The number of components which must be an integer}

\item{maxit}{The maximum number of iterations to terminate the algorithm. This must be an integer.}

\item{tol}{This is the tolenrance level of convergence dafault to 1e-10.}

\item{show_table}{This is a logical parameter which is default to FALSE.}
}
\value{
res which is a list containing the component mean and covariance matrix, the information criterion
}
\description{
zipCWM Function for the clustering of data
}
\keyword{Algorithm}
\keyword{Clustering}
\keyword{MzipCWM}

